pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.6.2",
    "click>=8.3.1",
    "dj-database-url>=3.1.2",
    "django>=5.2.12",
    "django-celery-beat>=2.9.0",
    "django-htmx>=1.27.0",
    "httpx>=0.28.1",
    "langgraph>=1.1.2",
    "litellm>=1.82.4",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.3",
    "python-decouple>=3.8",
    "python-telegram-bot>=22.7",
    "pyyaml>=6.0.3",
    "redis>=6.4.0",
]
